Rename : CalculatorFragment, ViewModel -> SearchFragment, ViewModel로 수정
