+++ app/src/main/java/sang/gondroid/calingredientfood/presentation/search/SearchViewModel.kt
@@ -1,4 +1,4 @@
-package sang.gondroid.calingredientfood.presentation.calculator
+package sang.gondroid.calingredientfood.presentation.search
 
 import android.view.View
 import sang.gondroid.calingredientfood.presentation.base.BaseViewModel
@@ -12,19 +12,18 @@ import sang.gondroid.calingredientfood.domain.model.FoodNtrIrdntModel
 import sang.gondroid.calingredientfood.domain.model.Model
 import sang.gondroid.calingredientfood.domain.use_case.GetFoodNtrIrdntListUseCase
 import sang.gondroid.calingredientfood.domain.util.ViewType
-import sang.gondroid.calingredientfood.presentation.util.DebugLog
 import sang.gondroid.calingredientfood.presentation.util.SearchMode
 import sang.gondroid.calingredientfood.presentation.util.UIState
 
-internal class CalculatorViewModel(
+internal class SearchViewModel(
     private val getFoodNtrIrdntUseCase: GetFoodNtrIrdntListUseCase
 ) : BaseViewModel() {
 
     /**
      * Gon [22.02.22] : 1. Spinner 선택된 아이템(SearchMode)을 이용해 관리되는 LiveData
-     *                     fragment_calculator.xml calculator_search_editText Hint 변경에 사용
+     *                     fragment_search.xml search_editText Hint 변경에 사용
      *
-     * Gon [22.02.22] : 2, 3. LiveData를 이용해 값이 변경되면 fragment_calculator.xml의 표현식을 통해 BindingAdapter 호출
+     * Gon [22.02.22] : 2, 3. LiveData를 이용해 값이 변경되면 fragment_search.xml의 표현식을 통해 BindingAdapter 호출
      *                        UIState.Success 인 경우 BindingAdapter.submitList() 메서드가 호출됨
      */
     private val _currentSearchModeLiveData: MutableLiveData<SearchMode> = MutableLiveData(SearchMode.FOOD)
