+++ app/src/main/java/sang/gondroid/calingredientfood/presentation/util/Constants.kt
@@ -5,20 +5,4 @@ object Constants {
      * Gon [22.01.25] : FragmentManager에 해당 상수를 TAG로 BottomSheetDialogFragment를 추가하여 표시하기 위해 사용되는 상수 정의
      */
     const val BOTTOM_SHEET_TAG : String = "FOOD_NTR_IRDNT_BOTTOM_SHEET"
-
-    /**
-     * Gon [22.01.25] : SaveStateHandle의 Key값으로 사용되는 상수 정의 (검색모드)
-     */
-    const val SEARCH_MODE_KEY : String = "CURRENT_SEARCH_MODE"
-
-    /**
-     * Gon [22.01.25] : SaveStateHandle의 Key값으로 사용되는 상수 정의 (식품영양정보)
-     */
-    const val FOOD_NTR_IRDNT_LIST_KEY : String = "SELLECT_FOOD_NTR_IRDNT_LIST"
-
-    /**
-     * Gon [22.02.04] : SaveStateHandle의 Key값으로 사용되는 상수 정의 (식품영양 계산정보)
-     */
-    const val CALCULATOR_LIST_KEY : String = "SELLECT_CALCULATOR_LIST"
-
 }
