+++ app/src/test/java/sang/gondroid/calingredientfood/di/AppTestModule.kt
@@ -5,12 +5,12 @@ import org.koin.dsl.module
 import sang.gondroid.calingredientfood.data.repository.TestFoodNtrIrdntRepositoryImpl
 import sang.gondroid.calingredientfood.domain.repository.FoodNtrIrdntRepository
 import sang.gondroid.calingredientfood.domain.use_case.GetFoodNtrIrdntListUseCase
-import sang.gondroid.calingredientfood.presentation.calculator.CalculatorViewModel
+import sang.gondroid.calingredientfood.presentation.search.SearchViewModel
 
 internal val appTestModule = module {
 
     // ViewMdoel
-    viewModel { CalculatorViewModel(get()) }
+    viewModel { SearchViewModel(get()) }
 
     // UseCase
     factory { GetFoodNtrIrdntListUseCase(get<FoodNtrIrdntRepository>()) }
