+++ app/src/test/java/sang/gondroid/calingredientfood/viewmodel/TestSearchViewModel.kt
@@ -3,14 +3,14 @@ package sang.gondroid.calingredientfood.viewmodel
 import kotlinx.coroutines.test.runBlockingTest
 import org.junit.Test
 import org.koin.test.inject
-import sang.gondroid.calingredientfood.presentation.calculator.CalculatorViewModel
+import sang.gondroid.calingredientfood.presentation.search.SearchViewModel
 
 /**
- * Gon [22.02.22] : CalculatorViewModel을 주입받아 테스트
+ * Gon [22.02.22] : SearchViewModel을 주입받아 테스트
  *                  @ExperimentalCoroutinesApi : Coroutine Test가 실험용 라이브러리이기 때문에 추가
  */
-class TestCalculatorViewModel : ViewModelTest() {
-    private val viewModel: CalculatorViewModel by inject()
+class TestSearchViewModel : ViewModelTest() {
+    private val viewModel: SearchViewModel by inject()
 
     @Test
     fun `test viewModel searchFunc Success`() : Unit = runBlockingTest {
